+++ examples/resources/elasticstack_kibana_slo/resource.tf
@@ -133,6 +133,7 @@ resource "elasticstack_kibana_slo" "custom_histogram" {
     timeslice_window = "5m"
   }
 
+  tags = ["tag-1", "another_tag"]
 }
 
 //Available from 8.10.0