+++ internal/elasticsearch/cluster/slm_test.go
@@ -18,7 +18,7 @@ func TestAccResourceSLM(t *testing.T) {
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { acctest.PreCheck(t) },
 		CheckDestroy:             checkSlmDestroy(name),
-		ProtoV5ProviderFactories: acctest.Providers,
+		ProtoV6ProviderFactories: acctest.Providers,
 		Steps: []resource.TestStep{
 			{
 				Config: testAccSlmCreate(name),