+++ internal/elasticsearch/enrich/policy_test.go
@@ -18,7 +18,7 @@ func TestAccResourceEnrichPolicy(t *testing.T) {
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { acctest.PreCheck(t) },
 		CheckDestroy:             checkEnrichPolicyDestroy(name),
-		ProtoV5ProviderFactories: acctest.Providers,
+		ProtoV6ProviderFactories: acctest.Providers,
 		Steps: []resource.TestStep{
 			{
 				Config: testAccEnrichPolicy(name),