+++ internal/elasticsearch/index/data_stream_test.go
@@ -18,7 +18,7 @@ func TestAccResourceDataStream(t *testing.T) {
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { acctest.PreCheck(t) },
 		CheckDestroy:             checkResourceDataStreamDestroy,
-		ProtoV5ProviderFactories: acctest.Providers,
+		ProtoV6ProviderFactories: acctest.Providers,
 		Steps: []resource.TestStep{
 			{
 				Config: testAccResourceDataStreamCreate(dsName),