+++ internal/elasticsearch/ingest/processor_lowercase_data_source_test.go
@@ -10,7 +10,7 @@ import (
 func TestAccDataSourceIngestProcessorLowercase(t *testing.T) {
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { acctest.PreCheck(t) },
-		ProtoV5ProviderFactories: acctest.Providers,
+		ProtoV6ProviderFactories: acctest.Providers,
 		Steps: []resource.TestStep{
 			{
 				Config: testAccDataSourceIngestProcessorLowercase,