+++ internal/elasticsearch/security/role_mapping_test.go
@@ -18,7 +18,7 @@ func TestResourceRoleMapping(t *testing.T) {
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { acctest.PreCheck(t) },
 		CheckDestroy:             checkResourceSecurityRoleMappingDestroy,
-		ProtoV5ProviderFactories: acctest.Providers,
+		ProtoV6ProviderFactories: acctest.Providers,
 		Steps: []resource.TestStep{
 			{
 				Config: testAccResourceSecurityRoleMappingCreate(roleMappingName),