+++ internal/fleet/enrollment_tokens_data_source_test.go
@@ -14,7 +14,7 @@ var minVersionEnrollmentTokens = version.Must(version.NewVersion("8.6.0"))
 func TestAccDataSourceEnrollmentTokens(t *testing.T) {
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { acctest.PreCheck(t) },
-		ProtoV5ProviderFactories: acctest.Providers,
+		ProtoV6ProviderFactories: acctest.Providers,
 		CheckDestroy:             checkResourceAgentPolicyDestroy,
 		Steps: []resource.TestStep{
 			{